express offsets as float64 so that we can use sort.Float64s (see newer commits)

+++ client/model/chord.go
@@ -47,7 +47,7 @@ func (chord Chord) updateScore(score *Score) error {
 		}
 
 		part.LastOffset = part.CurrentOffset
-		part.CurrentOffset += float32(shortestDurationMs)
+		part.CurrentOffset += shortestDurationMs
 	}
 
 	return nil
